+++ tests/test_admin.py
@@ -1,10 +1,10 @@
-from unittest import skip
+from unittest import skipUnless
 
 from .base import BaseUrlTestCase
 
 
 class UrlManagerContentTypeSearchTestCase(BaseUrlTestCase):
-    @skip("Failed test should be addresses in future ticket")
+
     def test_get_search_results(self):
         """
         A filtered queryset is returned containing matches with a given search_term
@@ -12,6 +12,12 @@ class UrlManagerContentTypeSearchTestCase(BaseUrlTestCase):
         """
         self.url2.content_object = self.page2
         self.url2.save()
+
+        url2_version = self.url2.versions.last()
+        url2_version.publish(self.superuser)
+
+        url1_version = self.url.versions.first()
+        url1_version.publish(self.superuser)
 
         search_term = self.page.get_title()
         results, use_distinct = self.url_admin.get_search_results(
@@ -22,7 +28,6 @@ class UrlManagerContentTypeSearchTestCase(BaseUrlTestCase):
         self.assertEqual(results.last(), self.url2)
         self.assertEqual(results.count(), 2)
 
-    @skip("Failed test should be addresses in future ticket")
     def test_get_search_results_partial_search_term(self):
         """
         A filtered queryset is returned containing matches with a given search_term
@@ -41,10 +46,9 @@ class UrlManagerContentTypeSearchTestCase(BaseUrlTestCase):
         self.assertEqual(results.last(), self.url2)
         self.assertEqual(results.count(), 2)
 
-    # @skipUnless(
-    #     BaseUrlTestCase.is_versioning_enabled(), "Test only relevant for versioning"
-    # )
-    @skip("Failed test should be addresses in future ticket")
+    @skipUnless(
+        BaseUrlTestCase.is_versioning_enabled(), "Test only relevant for versioning"
+    )
     def test_get_search_results_versioning(self):
         from djangocms_versioning.constants import DRAFT, PUBLISHED