+++ tests/test_templatetags.py
@@ -7,10 +7,11 @@ from .base import BaseUrlTestCase
 
 class UrlManagerTemplateTagsTestCase(BaseUrlTestCase):
     url_template = (
-        """{% load djangocms_url_manager_tags %}{% render_url url %}"""
+        """{% load djangocms_url_manager_tags %}{% render_url url.url_grouper %}"""
     )  # noqa: E501
 
     def test_render_url(self):
+        self.url.versions.first().publish(user=self.superuser)
         output = self.render_template_obj(
             self.url_template, {"url": self.url}, self.get_request("/")
         )
@@ -19,6 +20,7 @@ class UrlManagerTemplateTagsTestCase(BaseUrlTestCase):
         self.assertEqual(parsed.path, "/en/test/")
 
     def test_render_url_other_site(self):
+        self.url.versions.first().publish(user=self.superuser)
         self._create_url_override(self.url, self.site2, self.page2)
         with override_settings(SITE_ID=self.site2.pk):
             output = self.render_template_obj(