+++ tests/test_views.py
@@ -1,10 +1,13 @@
-from unittest import skip, skipIf, skipUnless
+from unittest import skipIf, skipUnless
 
 from django.contrib.contenttypes.models import ContentType
 
 from cms.models import Page, User
+from cms.utils.urlutils import admin_reverse
 
 from djangocms_url_manager.compat import CMS_36
+from djangocms_url_manager.constants import SELECT2_URLS
+from djangocms_url_manager.utils import is_versioning_enabled
 
 from .base import BaseUrlTestCase
 
@@ -67,6 +70,33 @@ class UrlManagerSelect2ContentObjectViewsTestCase(BaseUrlTestCase):
                 {"text": self.page.get_title(), "id": self.page.pk},
                 {"text": self.page2.get_title(), "id": self.page2.pk},
             ],
+        )
+
+    def test_select2_view(self):
+
+        with self.login_user_context(self.superuser):
+            response = self.client.get(
+                admin_reverse(
+                    SELECT2_URLS,
+                ),
+            )
+
+        result = [self.url.pk, self.url2.pk, ]
+        text_result = []
+
+        if is_versioning_enabled():
+            # The following versions have draft content
+            text_result.append(f"{self.url.internal_name} (Not published)")
+            text_result.append(f'{self.url2.internal_name} (Not published)')
+        else:
+            text_result.append(f"{self.url.internal_name}")
+            text_result.append(f"{self.url2.internal_name}")
+
+        self.assertEqual(response.status_code, 200)
+        self.assertEqual([a['id'] for a in response.json()['results']], result)
+        self.assertEqual(
+            [a['text'] for a in response.json()['results']],
+            text_result,
         )
 
     def test_return_poll_content_in_select2_view(self):
@@ -219,8 +249,12 @@ class UrlManagerSelect2UrlsViewsTestCase(BaseUrlTestCase):
         response = self.client.get(self.select2_urls_endpoint)
         self.assertEqual(response.status_code, 403)
 
-    @skip("Failed test should be addresses in future ticket")
     def test_select2_url_view_without_site_id(self):
+        url1_version = self.url.versions.first()
+        url1_version.publish(self.superuser)
+        url2_version = self.url2.versions.last()
+        url2_version.publish(self.superuser)
+
         with self.login_user_context(self.superuser):
             response = self.client.get(self.select2_urls_endpoint)
         self.assertEqual(response.status_code, 200)
@@ -228,8 +262,10 @@ class UrlManagerSelect2UrlsViewsTestCase(BaseUrlTestCase):
             [p["id"] for p in response.json()["results"]], [self.url.pk, self.url2.pk]
         )
 
-    @skip("Failed test should be addresses in future ticket")
     def test_select2_url_view_with_site_id(self):
+        url2_version = self.url2.versions.first()
+        url2_version.publish(self.superuser)
+
         with self.login_user_context(self.superuser):
             response = self.client.get(
                 self.select2_urls_endpoint, data={"site": self.site2.pk}
@@ -237,21 +273,28 @@ class UrlManagerSelect2UrlsViewsTestCase(BaseUrlTestCase):
         self.assertEqual(response.status_code, 200)
         self.assertEqual([p["id"] for p in response.json()["results"]], [self.url2.pk])
 
-    @skip("Failed test should be addresses in future ticket")
     def test_select2_url_view_with_object_pk(self):
+        url_version = self.url.versions.first()
+        url_version.publish(self.superuser)
+
         with self.login_user_context(self.superuser):
             response = self.client.get(
-                self.select2_urls_endpoint, data={"pk": self.url2.pk}
+                self.select2_urls_endpoint, data={"pk": self.url.pk}
             )
         self.assertEqual(response.status_code, 200)
-        self.assertEqual([a["id"] for a in response.json()["results"]], [self.url2.pk])
+        self.assertEqual([a["id"] for a in response.json()["results"]], [self.url.pk])
 
-    @skip("Failed test should be addresses in future ticket")
     def test_select2_url_view_set_limit(self):
-        self._create_url(anchor="test")
+        url = self.url.versions.first()
+        url.publish(self.superuser)
+
+        url2_version = self.url2.versions.first()
+        url2_version.publish(self.superuser)
+
         with self.login_user_context(self.superuser):
-            response = self.client.get(self.select2_urls_endpoint, data={"limit": 2})
+            response = self.client.get(self.select2_urls_endpoint, data={"limit": 1})
+
         content = response.json()
         self.assertEqual(response.status_code, 200)
         self.assertTrue(content["more"])
-        self.assertEqual(len(content["results"]), 2)
+        self.assertEqual(len(content["results"]), 1)